+++ cmd/collector/main.go
@@ -25,7 +25,12 @@ import (
 	"firewall-collector/internal/tftp"
 )
 
-const version = "1.2.94"
+const version = "1.2.95"
+
+
+
+
+
 
 type Collector struct {
 	cfg            *config.ProbeConfig